[BUGFIX] Clear autoloader cache if dbal is loaded in 1-2-3
The 1-2-3 install tool loads dbal on demand if some php extensions like pdo is
loaded. It adds dbal to extList and forces a reload. The core autoloader
creates a cache entry depending on the current loaded extensions. So the cache
must be cleared before redirect to force a recalculation of autoloadable classes.
This prevents a fatal error 'class tx_dbal_querycache not found' during install.

Change-Id: I7ec560e8639b3e95ec78739fde049d508f3e3d29
Resolves: #28187
Related: #28186
Related: #28063