[TASK] Make Core\Configuration\ConfigurationManager non static
To have the possibility to remove dependencies from code that 
calls the configuration manager, it has to be made non static.

Together with this change, all further configuration file logic
is moved to that class, to not expose internals in other
places. For that some methods needed to be made public, but
have been marked private, thus are not public API.

Further cleanup and simplification, especially not calling
makeInstance several times, will go in a separate change.

Change-Id: I18392f2b222309e58f1addc8b6ec8eb54f0d78f5
Resolves: #41052
Releases: 6.0
Reviewed-on: http://review.typo3.org/15268
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel