[TASK] Only use one select to calculate page cache timeout
To get the first time value we do 2 SQL select statements at the
moment but this can be combined in one SQL select statement. This
will save as time and reduces bandwidth to/from the SQL server.

The one SQL select statement from 4.6 had a bug and was changed
to use 2 SQL select statements instead of fixing the statement itself.
So this is a revert to the old source with the fixed statement.

Resolves: #38838
Releases: master
Change-Id: Ida3542ef31e696a9457042c5980e9aa0c6cdfede
Reviewed-on: http://review.typo3.org/43613
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>