[BUGFIX] exec_INSERTmultipleRows doesn't check alternative handlers
The function just passes the function call to it's parents insert
function when the type of the current handler is "native".
When using an alternative handler than "_DEFAULT" for a table, wich
connection is also "native", it will directly forward the call without
respecting my mapping for the table to another handler.

Similar to the other methods, this is solved by first fetching
the handler for the current table and then feeding the created
SQL string to mysql_query on given handler link.

The patch additionaly fixes db errors in 6.1 after mysqli switch.

Change-Id: I0c984b355916a99603ed72f0606e173608e4de81
Fixes: #34327
Releases: 6.1, 6.0, 4.7, 4.5
Reviewed-on: https://review.typo3.org/9272
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn