[TASK] Remove unused hook ExtensionManagerHooks
With the rewrite of the extension manager in 6.0, a hook of dbal in
the old extension manager is unused and rotting as dead code in dbal.

The interface that is implemented does not exist anymore.

The hook is removed for now, there is currently no active
contribution to handle this specific piece of code. If it is
introduced later again, it needs to be adapted to the new extension
manager, so the old hook could only be used as basic idea, but not
as production code at the moment.

Change-Id: Ie15962ba865b1be676c93be39228a7e9c25bd683
Resolves: #47762
Releases: 6.2
Reviewed-on: https://review.typo3.org/20365
Reviewed-by: Thomas Maroschik
Tested-by: Thomas Maroschik
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn