[TASK] Deprecate path prefixes in callUserFunction and getUserObj
The possibility to call "GeneralUtility::callUserFunc()"
and "getUserObj()" with the first parameter containing the file reference
to the function or class separated by a colon is deprecated.

An example is "EXT:realurl/class.tx_realurl.php:tx_realurl->encodeSpURL".

Calling one of the methods above with this syntax now triggers a
deprecation message.

It is encouraged to use the common autoloading functionality of
composer or via the fallback autoloader by TYPO3 to achieve the
same functionality.

Resolves: #76345
Releases: master
Change-Id: Iaaa2f825d330681199f0b0928727b93bbd3c73a7
Reviewed-on: https://review.typo3.org/48352
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>