[BUGFIX] Ensure correct mysqli resource for admin_get_tables query
The admin_get_tables method tries to fetch all available database tables
from the default connection. But using the native mode, the query is
executed with the $this->lastHandlerKey link which might not be the
default connection. This patch ensures the correct mysqli resource is
used to resolve the database queries.

See: https://forge.typo3.org/issues/80358

Resolves: #1